+++ somme.c
@@ -5,24 +5,24 @@
 #include "calculate.h"
 
 
-
 void somme(void) {
-	
-	int nombre1 = 0;
-	int nombre2 = 0;
-	int resultat_nombre;
+
+
+	double nombre1 = 0;
+	double nombre2 = 0;
+	double resultat_nombre;
 	int choix = 0;
 	int continuer = 1;
 
 	printf("\nEntrez le premier nombre:\n");
-	scanf("%d", &nombre1);
+	scanf("%lf", &nombre1);
 
 	printf("Entrez le deuxième nombre:\n");
-	scanf("%d", &nombre2);
+	scanf("%lf", &nombre2);
 
 	resultat_nombre = nombre1 + nombre2;
 
-	printf("\n\nLe résultat de %d + %d est %d\n", nombre1, nombre2, resultat_nombre);
+	printf("\n\nLe résultat de %.3lf + %.3lf est %.3lf\n", nombre1, nombre2, resultat_nombre);
 
 	printf("1/ Retour\n");
 	printf("2/ Quitter\n");
@@ -45,4 +45,5 @@ void somme(void) {
 			printf("Saisie incorrecte...\n");
 		}
 	}
+
 }
